webapp/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
xrpl-py==2.6.0
pydantic==2.10.3
python-dotenv==1.0.0
python-multipart==0.0.9
